update coupled scattering matrix so that interaction objects are loaded inside the matrix

+++ src/algebra/SMatrix.h
@@ -21,7 +21,7 @@ class SerialMatrix {
   /// Class variables
   int numRows_;
   int numCols_;
-  int numElements_;
+  size_t numElements_;
 
   T* mat = nullptr;  // pointer to the internal array structure.
 
@@ -93,7 +93,7 @@ class SerialMatrix {
   int localCols() const;
   /** Find global number of matrix elements
    */
-  int size() const;
+  size_t size() const;
   /** Get and set operator
    */
   T& operator()(const int &row, const int &col);
@@ -128,7 +128,7 @@ class SerialMatrix {
     if(numRows_ != m1.rows() || numCols_ != m1.cols()) {
       Error("Cannot add matrices of different sizes.");
     }
-    for (int s = 0; s < size(); s++) mat[s] += m1.mat[s];
+    for (size_t s = 0; s < size(); s++) mat[s] += m1.mat[s];
     return *this;
   }
 
@@ -138,21 +138,21 @@ class SerialMatrix {
     if(numRows_ != m1.rows() || numCols_ != m1.cols()) {
       Error("Cannot subtract matrices of different sizes.");
     }
-    for (int s = 0; s < size(); s++) mat[s] -= m1.mat[s];
+    for (size_t s = 0; s < size(); s++) mat[s] -= m1.mat[s];
     return *this;
   }
 
   /** Matrix-scalar multiplication.
    */
   SerialMatrix<T> operator*=(const T& that) {
-    for (int s = 0; s < size(); s++) mat[s] *= that;
+    for (size_t s = 0; s < size(); s++) mat[s] *= that;
     return *this;
   }
 
   /** Matrix-scalar division.
    */
   SerialMatrix<T> operator/=(const T& that) {
-    for (int s = 0; s < size(); s++) mat[s] /= that;
+    for (size_t s = 0; s < size(); s++) mat[s] /= that;
     return *this;
   }
 
@@ -213,7 +213,7 @@ SerialMatrix<T>::SerialMatrix(const int& numRows, const int& numCols,
   numCols_ = numCols;
   numElements_ = numRows_ * numCols_;
   mat = new T[numRows_ * numCols_];
-  for (int i = 0; i < numElements_; i++) mat[i] = 0;  // fill with zeroes
+  for (size_t i = 0; i < numElements_; i++) mat[i] = 0;  // fill with zeroes
   assert(mat != nullptr);  // Memory could not be allocated, end program
 }
 
@@ -238,7 +238,7 @@ SerialMatrix<T>::SerialMatrix(const SerialMatrix<T>& that) {
   }
   mat = new T[numElements_];
   assert(mat != nullptr);
-  for (int i = 0; i < numElements_; i++) {
+  for (size_t i = 0; i < numElements_; i++) {
     mat[i] = that.mat[i];
   }
 }
@@ -286,7 +286,7 @@ int SerialMatrix<T>::localCols() const {
   return numCols_;
 }
 template <typename T>
-int SerialMatrix<T>::size() const {
+size_t SerialMatrix<T>::size() const {
   return numElements_;
 }
 
@@ -336,7 +336,7 @@ int SerialMatrix<T>::global2Local(const int& row, const int& col) {
 template <typename T>
 std::vector<std::tuple<int, int>> SerialMatrix<T>::getAllLocalStates() {
   std::vector<std::tuple<int, int>> x;
-  for (int k = 0; k < numElements_; k++) {
+  for (size_t k = 0; k < numElements_; k++) {
     std::tuple<int, int> t = local2Global(k);  // bloch indices
     x.push_back(t);
   }
@@ -387,7 +387,7 @@ double SerialMatrix<T>::squaredNorm() {
 template <typename T>
 T SerialMatrix<T>::dot(const SerialMatrix<T>& that) {
   T scalar = (T)0.;
-  for (int i = 0; i < numElements_; i++) {
+  for (size_t i = 0; i < numElements_; i++) {
     scalar += (*(mat + i)) * (*(that.mat + i));
   }
   return scalar;
